Version 1.2.1
Added a List<> to keep the order of [BuildingTypes] since it must keep the same order of insertion. See wiki for more discussion.
This change will bring more difficulties when we implement "adding to a category" functionality. We have changed the position of BuildingTypes in GlobalProperty.SublistIndex to save execution time.

And some minor fixes.